+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 修复领地重新选区购买表单价格显示错误 @engsr6982
 - 修复未设置领地传送点进行跨纬度传送失败 #106 @engsr6982
+- 修复领地最小间距未生效问题 #115 @engsr6982
 
 ### ✨ 新增功能
 
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 实现脏数据计数 #95 @engsr6982
 - 重构经济系统，LegacyMoney 更改为可选依赖 @engsr6982
 - 选择领地表单支持**分页、搜索、过滤** #107 #101 @engsr6982
+- 新增配置项 `land.minSpacingIncludeY` #115 @engsr6982
 
 ### 🧹 其他改动